# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardCredentials {
  username: string;
  password: string;
}

export type DashboardUserRole =
  (typeof DashboardUserRole)[keyof typeof DashboardUserRole];

export const DashboardUserRole = {
  admin: "admin",
  employee: "employee",
} as const;

export interface DashboardUser {
  id: number;
  username: string;
  role: DashboardUserRole;
}

export interface OrderItemSelectedOption {
  groupName: string;
  choice: string;
}

export interface OrderItemCustomization {
  size?: string;
  riceType?: string;
  addon?: string;
  selectedOptions?: OrderItemSelectedOption[];
}

export interface OrderItem {
  id: string;
  name: string;
  price: number;
  quantity: number;
  customization?: OrderItemCustomization;
}

export type OrderOrderType =
  (typeof OrderOrderType)[keyof typeof OrderOrderType];

export const OrderOrderType = {
  delivery: "delivery",
  pickup: "pickup",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  done: "done",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  dailyNumber: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  totalPrice: number;
  deliveryFee: number;
  discountCode?: string | null;
  discountAmount?: number | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  paymentMethod: string;
  notes?: string | null;
  createdAt: string;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  done: "done",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface RevenueAggregate {
  totalRevenue: number;
  deliveryRevenue: number;
  itemsRevenue: number;
  orderCount: number;
  taxAmount: number;
  netRevenue: number;
  cancelledCount: number;
  cancelledValue: number;
  pendingCount: number;
  cashCount: number;
  onlineCount: number;
  cashRevenue: number;
  onlineRevenue: number;
}

export interface DailyBreakdown {
  date: string;
  total: number;
  delivery: number;
  items: number;
  orders: number;
  tax: number;
  net: number;
  cancelledCount: number;
  cancelledValue: number;
  cashCount: number;
  onlineCount: number;
}

export interface MonthlyBreakdown {
  month: string;
  total: number;
  delivery: number;
  items: number;
  orders: number;
  tax: number;
  net: number;
  cancelledCount: number;
  cancelledValue: number;
  cashCount: number;
  onlineCount: number;
}

export interface TopItem {
  id: string;
  name: string;
  qty: number;
  revenue: number;
}

export interface LiveRevenueData {
  lastHour: RevenueAggregate;
  last30min: RevenueAggregate;
  uniqueCustomerCount: number;
  totalItemsSold: number;
  totalDiscounts: number;
}

export interface RevenueData {
  today: RevenueAggregate;
  week: RevenueAggregate;
  month: RevenueAggregate;
  year: RevenueAggregate;
  dailyBreakdown: DailyBreakdown[];
  monthlyBreakdown: MonthlyBreakdown[];
  topItems: TopItem[];
}

export interface SizeOption {
  name: string;
  price: number;
  enabled: boolean;
}

export interface MenuItem {
  id: number;
  itemId: string;
  name: string;
  nameEn?: string | null;
  category: string;
  price: number;
  available: boolean;
  imageKey?: string | null;
  imageUrl?: string | null;
  stock?: number | null;
  sizes: SizeOption[];
  sortOrder: number;
}

export interface MenuItemUpdate {
  name?: string;
  nameEn?: string;
  price?: number;
  available?: boolean;
  stock?: number | null;
  imageUrl?: string | null;
  imageKey?: string | null;
  sizes?: SizeOption[];
}

export interface Driver {
  id: number;
  name: string;
  phone: string;
  active: boolean;
  photoUrl?: string | null;
}

export interface DriverDailySummary {
  driverId: number;
  name: string;
  orderCount: number;
  cashTotal: number;
  electronicTotal: number;
}

export interface DriverRatingInput {
  /**
   * @minimum 1
   * @maximum 5
   */
  stars: number;
  comment?: string | null;
}

export interface DriverReview {
  orderId: number;
  stars: number;
  comment?: string | null;
  createdAt?: string;
}

export interface DriverRankingEntry {
  id: number;
  name: string;
  phone: string;
  photoUrl?: string | null;
  active: boolean;
  completedDeliveries: number;
  totalRatings: number;
  avgStars?: number | null;
  recentReviews: DriverReview[];
}

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  phone?: string;
  search?: string;
  from?: string;
  to?: string;
  limit?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  done: "done",
  cancelled: "cancelled",
} as const;

export type GetRevenueRangeParams = {
  from: string;
  to: string;
};

export type SubmitDriverRating200 = {
  ok?: boolean;
  stars?: number;
};