# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, preparing, ready, done, cancelled]
        - name: phone
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Revenue / KPIs ───────────────────────────────────────────────────
...
components:
  schemas:
...
    OrderItemSelectedOption:
      type: object
      properties:
        groupName:
          type: string
        choice:
          type: string
      required: [groupName, choice]
    OrderItemCustomization:
      type: object
      properties:
        size:
          type: string
        riceType:
          type: string
        addon:
          type: string
        selectedOptions:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemSelectedOption"
    OrderItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: integer
        quantity:
          type: integer
        customization:
          $ref: "#/components/schemas/OrderItemCustomization"
      required: [id, name, price, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        dailyNumber:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: integer
        deliveryFee:
          type: integer
        discountCode:
          type: string
          nullable: true
        discountAmount:
          type: integer
          nullable: true
        orderType:
          type: string
          enum: [delivery, pickup]
        status:
          type: string
          enum: [pending, preparing, ready, done, cancelled]
        paymentMethod:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - dailyNumber
        - customerName
        - customerPhone
        - customerAddress
        - items
        - totalPrice
        - deliveryFee
        - orderType
        - status
        - paymentMethod
        - createdAt
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, done, cancelled]
      required: [status]

    # ── Revenue ───────────────────────────────────────────────────────
...