# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemCustomization } from "./orderItemCustomization";

export interface OrderItem {
  id: string;
  name: string;
  price: number;
  quantity: number;
  customization?: OrderItemCustomization;
}