# lib/api-zod/src/generated/types/orderItemCustomization.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemSelectedOption } from "./orderItemSelectedOption";

export interface OrderItemCustomization {
  size?: string;
  riceType?: string;
  addon?: string;
  selectedOptions?: OrderItemSelectedOption[];
}